+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Change Log
 
 ### Changed
 - Kotlin updated to [v1.3.71](https://github.com/JetBrains/kotlin/releases/tag/v1.3.71)
+- OpenApi Codegen updated to [v4.3.0](https://github.com/OpenAPITools/openapi-generator/releases/tag/v4.3.0)
 
 ## 2.0.1 (2020-03-10)