+++ CHANGELOG.md
@@ -3,8 +3,19 @@ Change Log
 
 ## 2.1.0 (TBD)
 
+### Added
+- @Path params are always first in operation.
+- Option to force variables of Composed schema (oneOf and anyOf) as not required (nullable).
+
 ### Changed
 - Kotlin updated to [v1.3.71](https://github.com/JetBrains/kotlin/releases/tag/v1.3.71)
+- OpenApi Codegen updated to [v4.3.0](https://github.com/OpenAPITools/openapi-generator/releases/tag/v4.3.0)
+
+### Fixed
+- Kotlin allOf inheritance.
+
+### Removed
+- Not used constants.
 
 ## 2.0.1 (2020-03-10)